message: fix error counts
Previous versions of Message would accept a raw bit error count
from the Framer. The error count would include the "garbage"
characters at the end of message that are read before the framer
can decide where the end-of-message is.

Instead of accepting a count of errors, accept a u8 slice which
counts the errors in each individual byte. The errors are now
tallied correctly.