equalize: add adaptive equalizer
The Equalizer can remove ISI due to frequency-selective effects in
the channel. It accepts byte-aligned samples (at two samples/sym)
and outputs estimated bytes.

FM tends to attenuate high frequencies more than low frequencies.
Some receivers will try to correct for this with
pre-emphasis/de-emphasis filters… others, not so much. The
adaptive equalizer relieves us of the burden of accounting for
this.