message, framing: assemble bytes into messages
Add the Layer 2 protocol decoding for SAME/EAS… such that it is.
Each individual data burst is collected into a Burst. We use a
sliding window of the past three bursts to try and frame a
message.

Errors are corrected with 2-of-3 columnar parity correction,
which is a really stupid way to correct errors.