squelch: output in chunks of 16 samples
Previous versions of the CodeCorrelator would always output an
iterator with 64 samples.

This is too many for the Equalizer, which expects inputs in chunks
of sixteen. The extra samples risk getting "stranded" if they are
not consumed before Framer returns an event of interest to the
caller.

Use an output array of 16 samples instead. This is a cleaner API
anyway.