codesquelch, framer: improve integration
Improve integration between the squelch and the framer by

* tracking symbol power; and
* capping the maximum inter-burst time

The CodeSquelch, renamed to CodeAndPowerSquelch, now squelches
with both power and code. The CodeAndPowerSquelch can enforce a
minimum power level to open or to remain open. The squelch must
track the power history in order to prevent it from closing
early.

If we don't receive a complete set of three SAME bursts, the old
bursts will hang around in the buffer forever. When we receive the
next burst, the parity correction will try and use it. This can
cause old messages to be emitted and/or for new messages to get
eaten. Not good.

Make the squelch block track the total number of symbols it has
received. This is useful as a measure of time. The Framer now
ingests this estimate and will purge old data if the
MAX_INTERBURST_GAP_SYMBOLS time elapses.

The squelch is now closed with end() instead of reset().