message: permit callsigns <8 characters
The sample SAME file on Wikipedia [1] uses a callsign of "WTSP/TV",
which is only seven characters long. The NWS spec [2] specifies
that callsigns are eight characters long… but they also give
example callsigns which are only seven.

The Wikipedia sample is a generated message and *not* a live
message. It is unclear what live stations actually do.

Loosen the parser to require anywhere from three to eight
characters for callsigns.

References
1. https://en.wikipedia.org/wiki/File:Same.wav
2. https://www.nws.noaa.gov/directives/sym/pd01017012curr.pdf