sameold: decode EOMs faster
When playing back or recording a SAME voice message, it is
sometimes desirable to terminate the playback as quickly as
possible. SAME voice messages themselves are unlikely to
nuisance-trip the decoder. Error correction for the
EOM—which carries no other data—is largely unnecessary.

Per the SAME spec [1 pp. 19], receivers are not required to wait
for three bursts of "`NNNN`" to end a message. One burst which
starts with "`NN`" is sufficient.

To improve standards-compliance, add dedicated decoding logic for
the EOM which is independent of the error-correction. Every
individual "`NNNN`" burst which decodes will output a separate
`Message::EndOfMessage`.

The `Message` parser now accepts two `N` characters as an EOM.
The framer must still decode all four `N` characters with no
greater than `max_prefix_bit_errors` in order to qualify as an EOM.

This is an API-breaking change.

References
1. https://www.nws.noaa.gov/directives/sym/pd01017012curr.pdf