sameold: add with_fast_end_of_message option
When playing back or recording a SAME voice message, it is
sometimes desirable to terminate the playback as quickly as
possible. SAME voice messages themselves are unlikely to
nuisance-trip the decoder. Error correction for the
EOM—which carries no other data—is largely unnecessary. We can
generally terminate recording as soon as we receive the first
EOM ("`NNNN`").

The `with_fast_end_of_message(true)` option causes the framer to
output a `Message::EndOfMessage` for every EOM burst received.

This is an API-expanding change.