sameold: log the full text of each SAME burst
Output an info-level debugging message with the contents of each
individual SAME burst, as they are received.

The burst text is guaranteed to be valid utf-8, and within the
maximum size limit for a SAME message, but absolutely *no* other
guarantees are made. Since SAME does not have an end-of-message
indicator, the log line may have trailing garbage characters.

Sample output (via `samedec -v`):

    INFO  sameold::framing     > burst: ended: "NNNN�UUU"