container: add containerized build and execution
In the Linux world, portability requires that we support very
old glibc. The Python `manylinux` environment currently requires
glibc >= 2.28, which supports el8 and later.

Github Runners are much more modern than this. If we want to
build against an older glibc, we'll need to do so in a container.

Thankfully, the Rust project provides offiical OCI images based on
slim-buster, which also has glibc 2.28. Add a Dockerfile which
builds and tests in slim-buster. We also support a static musl
target with alpine.

Our Dockerfile has a `localfile` target which can be used to
output a standalone binary instead of an image.