github: add release actions
We want to target multiple Linux architectures, but Github only
has amd64 runners. The best solution is to build a
*multi-architecture* Docker image.

We use pre-packaged Github actions to configure the kernel, qemu,
and Docker to work together to build and run non-native images.
We are not in the business of publishing container images (for
now). Instead, we use our "`localfile`" Dockerfile `--target` to
export the binary during the build process. glibc (2.28 for now)
and musl variants are available.

Builds for operating systems other than Linux take place on
Github Runner machines, without using Docker.

Releases trigger on tags and commits on `develop`, which are
turned into a "latest" release.