sameold: message: record voting byte count
Add `MessageHeader::voting_byte_count()` for counting the total
number of bytes which were examined with 2-of-3 bit voting. This
field, along with the error count, summarizes how well we believe
we decoded the message.

The new TryFrom conversion method also takes a byte slice for the
message contents directly instead of a utf-8 String.

This is an API-expanding change.