sameold: internal re-org, separate FrameOut
Re-organize all the receiver-specific functionality under
`mod receiver`. The `FrameOut` is part of the public-facing API,
so it is separated from the Framer and placed in a new
`crate::receiver::output` module.

In addition, `mod.rs` files are eliminated in favor of Rust 2018's
recommended layout.

These changes are internal only and do not affect API.