sameold: receiver: add new Assembler
The `SameReceiver`'s processing methods are split into two
parts:

1. The *Link Layer* handles demodulation and framing Bursts.

2. The *Transport Layer* builds individual Bursts into a
   single `Message` estimate. The transport layer also monitors
   for SAME voice messages that have exceeded their timeout.

The logic in the `Framer` which previously made `Message` esimates
is removed. Now the `Framer` only concerns itself with outputting
Bursts.

This is a MAJOR, API-BREAKING change.

* `FrameOut` is removed and replaced with `LinkOut`, which serves
  a similar purpose. `LinkOut` reports the link-layer state and
  includes a field for reporting raw Bursts as `Vec<u8>`.

* `SameReceiver::iter_frames()` is removed and replaced with
  `SameReceiver::iter_events()`, which serves a similar purpose.
  An "event" is a state change in either the link- or
  transport-layers. Events are reported via a new `SameEvent`
  type, which is `Display`-able.

* `SameReceiver::iter_messages()` works as it did before, but
  now it returns an opaque type.

The state reporting is further improved by modifying the
`codesquelch` to always output a state, even if no byte is ready.

Surprisingly, this does not break the API used by `samedec`, but
it does change behavior because `sameold` now suppresses duplicate
EOMs.