sameold,samedec: drop SignificanceLevel::Message
The `SignificanceLevel::Message` is not part of NWSI 10-1712 or
any other known standard, and very few event codes in sameold
ever implement it.

Discontinue this enum variant.

The following event codes are affected and are upgraded to
`Statement`s:

* ADR Administrative Message
* NMN Network Message Notification

This is an API-BREAKING change. The samedec CLI is unaffected, but
it is removed from the documentation.