sameold: update to strum 0.26
Thanks to changes in strum, we can no longer implement our own
`From<&str>` to choose what happens when the input is unmatched.
In order to keep the API, we must drop the `EnumString` derive
macro on `SignificanceLevel` and implement the parser ourselves.

Add `EnumIter` for use in unit tests.