sameold,samedec: report if message is national
Report valid national periodic tests and national emergencies
via `MessageHeader::is_national()`. Both the event and location
codes are checked.

These messages are reported to samedec child processes via the
`$SAMEDEC_IS_NATIONAL` environment variable, which is set to "`Y`"
for national messages and is empty otherwise.

This is an API-expanding and CLI-expanding change.