ci: use only one workflow everywhere
At least once in every developer's lifetime, you will craft the
*perfect* PR that passes all its tests with flying colors. You
click that big green merge button with complete confidence…

… only to have the post-merge "update the rolling release" job
fail hideously, requiring an immediate fix or revert. Blame is
passed. Ire is raised.

To minimize the chance of this happening, we reduce our CI jobs
down to a single workflow. We go ahead and run *all* the release
tasks, even for PRs, to make sure they will succeed if we run them
"for real."

The only jobs that are skipped are the ones which update tags and
publish new releases/packages. Everything else runs.