sameplace: split message parser into its own crate
The present `sameold` lib crate combines the digital modem with
a message text parser. This is not necessary.

* The interface from the modem to the text parser is just String.

* The modem depends on Message for its API, but it doesn't care
  about the finer details of the contents—viz., event codes.

To encourage re-use throughout the Rust ecosystem, we should
create crates which are narrow in scope.

Create a new crate for `Message` and friends. It includes the
text-parser without any of the modem DSP.

Documentation is given a touch-up.