sameold: remove slice conversion from Window
`SliceRingBuffer` is unsound and has multiple security
advisories [1] [2]. There has been no apparent progress on fixes
since May 2025 [3]. The crate's large volume of interdependent
`unsafe` blocks will likely make it very tricky to maintain.

While a memory-mapped deque is a very cool idea, it is time to
move on for now.

Let's replace `SliceRingBuffer` with a more traditional
double-ended queue.

These queues don't covert to slice, but we don't really *need* a
slice. At present, slice is mostly used to align the end of
`Window` (with the most recent samples) to the filter taps. We
can do that just as easily with a `DoubleEndedIterator`.

* Make `Window` convertible to an iterator of numbers, via
  `IntoIterator`. Replace all uses of slices with iterators.

* Remove slice conversions and accessors from Window. They're
  still technically accessible via `inner()`, but nothing uses
  them.

[1]: https://rustsec.org/advisories/RUSTSEC-2025-0044

[2]: https://rustsec.org/advisories/RUSTSEC-2020-0158

[3]: https://github.com/LiquidityC/slice_ring_buffer/issues/12