ci: make releases atomic
Previously, "build and test" jobs would upload releases themselves,
if they were cleared to do so. This mixing is not great from a
pipeline security standpoint since the entire workflow runs with
`contents: write`. It also introduces the potential for a
*partial release*. If some jobs complete, but others do not, then
a release will be issued without all its binaries.

Make this operation more atomic by performing publishing activities
in separate jobs. The publish jobs download all the artifacts and
create the release.